Removed Present function from SdJwt.Issuance
