+++ src/main/kotlin/eu/europa/ec/eudi/sdjwt/Disclosure.kt
@@ -148,12 +148,3 @@ sealed interface Disclosure {
         }
     }
 }
-
-/**
- * Concatenates the given disclosures into a single string, separated by
- * "~". The string also starts with "~".
- *
- * @receiver the disclosures to concatenate
- * @return the string as described above
- */
-fun Iterable<Disclosure>.concat(): String = fold("") { acc, disclosure -> "$acc~${disclosure.value}" }
