+++ src/main/kotlin/eu/europa/ec/eudi/sdjwt/Types.kt
@@ -120,66 +120,3 @@ sealed interface SdJwt<out JWT, out KB_JWT> {
         val keyBindingJwt: KB_JWT?,
     ) : SdJwt<JWT, KB_JWT>
 }
-
-/**
- * Creates a [presentation SD-JWT][SdJwt.Presentation]
- *
- * @param keyBindingJwt optional, the Holder Binding JWT to include
- * @param selectivelyDisclose a predicate of the [claims][SdJwt.Issuance.disclosures]
- * to be selectively disclosed into the presentation
- * @param JWT the type representing the JWT part of the SD-JWT
- * @param KB_JWT the type representing the Key Binding part of the SD-JWT
- * @return the presentation
- *
- * @receiver the [issued SD-JWT][SdJwt.Issuance] from which the presentation will be created
- */
-fun <JWT, KB_JWT> SdJwt.Issuance<JWT>.present(
-    keyBindingJwt: KB_JWT? = null,
-    selectivelyDisclose: (Claim) -> Boolean,
-): SdJwt.Presentation<JWT, KB_JWT> =
-    SdJwt.Presentation(
-        jwt,
-        disclosures.filter { disclosure ->
-            when (disclosure) {
-                is Disclosure.ArrayElement -> true // TODO Figure out what to do
-                is Disclosure.ObjectProperty -> selectivelyDisclose(disclosure.claim())
-            }
-        }.toSet(),
-        keyBindingJwt,
-    )
-
-/**
- * Serializes a [SdJwt.Issuance] to Combined Issuance Format
- *
- * @param serializeJwt a function to serialize the [JWT]
- * @param JWT the type representing the JWT part of the SD-JWT
- * @receiver the SD-JWT to serialize
- * @return the Combined Issuance format of the SD-JWT
- */
-fun <JWT> SdJwt.Issuance<JWT>.toCombinedIssuanceFormat(
-    serializeJwt: (JWT) -> String,
-): String {
-    val serializedJwt = serializeJwt(jwt)
-    val serializedDisclosures = disclosures.concat()
-    return "$serializedJwt$serializedDisclosures"
-}
-
-/**
- * Serialized a [SdJwt.Presentation] to Combined Presentation Format
- *
- * @param serializeJwt a function to serialize the [JWT]
- * @param JWT the type representing the JWT part of the SD-JWT
- * @param serializeKeyBindingJwt a function to serialize the [KB_JWT]
- * @param KB_JWT the type representing the Key Binding part of the SD
- * @receiver the SD-JWT to serialize
- * @return the Combined Presentation format of the SD-JWT
- */
-fun <JWT, KB_JWT> SdJwt.Presentation<JWT, KB_JWT>.toCombinedPresentationFormat(
-    serializeJwt: (JWT) -> String,
-    serializeKeyBindingJwt: (KB_JWT) -> String,
-): String {
-    val serializedJwt = serializeJwt(jwt)
-    val serializedDisclosures = disclosures.concat()
-    val serializedKbJwt = keyBindingJwt?.run(serializeKeyBindingJwt) ?: ""
-    return "$serializedJwt$serializedDisclosures~$serializedKbJwt"
-}
